+++ src/quantize/quantizer.rs
@@ -1,4 +1,3 @@
-use ahash::HashMap;
 use indexmap::IndexMap;
 
 use crate::color::Argb;
@@ -7,12 +6,12 @@ pub trait Quantizer {
     fn quantize(
         &mut self,
         pixels: &[Argb],
-        max_colors: i32,
+        max_colors: usize,
         return_input_pixel_to_cluster_pixel: Option<bool>,
     ) -> QuantizerResult;
 }
 
 pub struct QuantizerResult {
     pub color_to_count: IndexMap<Argb, u32>,
-    pub input_pixel_to_cluster_pixel: HashMap<Argb, Argb>,
+    pub input_pixel_to_cluster_pixel: IndexMap<Argb, Argb>,
 }